+++ server/src/main/java/org/opensearch/index/codec/AdditionalCodecs.java
@@ -9,7 +9,6 @@
 package org.opensearch.index.codec;
 
 import org.apache.lucene.codecs.Codec;
-import org.opensearch.OpenSearchException;
 import org.opensearch.common.Nullable;
 import org.opensearch.common.annotation.ExperimentalApi;
 import org.opensearch.index.IndexSettings;
@@ -19,7 +18,7 @@ import java.util.Map;
 import java.util.function.Supplier;
 
 /**
- * The {@link CodecRegistry} is designated as a mechanism to contribute additional {@link Codec}
+ * The {@link AdditionalCodecs} is designated as a mechanism to contribute additional {@link Codec}
  * that require non-trivial instantiation logic (could not be registered using service
  * loader).
  *
@@ -32,15 +31,16 @@ import java.util.function.Supplier;
  * @opensearch.experimental
  */
 @ExperimentalApi
-public interface CodecRegistry {
+public interface AdditionalCodecs {
     /**
      * Return an additional {@code Codec} instances to be registered with the default
-     * (or custom, if provided) {@link CodecService}. In case there is already registered
-     * {@code Codec} with the same name, the implementation calls {@code onConflict} method
-     * to resolve the conflict.
+     * (or custom, if provided) {@link CodecService}.
      *
      * <b>Note</b>: The default codec (see {@link CodecService#LUCENE_DEFAULT_CODEC}) could not be changed
-     * or removed.
+     * or removed, attempt to do so will lead to exception.
+     *
+     * <b>Note</b>: In case there is already registered {@code Codec} with the same name the exception
+     * will be raised.
      *
      * @param mapperService mapper service (if available)
      * @param indexSettings index settings
@@ -48,23 +48,4 @@ public interface CodecRegistry {
      * @return additional {@code Codec} instances, should not be "null"
      */
     Map<String, Codec> getCodecs(@Nullable MapperService mapperService, IndexSettings indexSettings, Supplier<Codec> defaultCodec);
-
-    /**
-     * In case there is already registered {@code Codec} with the same name, allows to resolve
-     * the conflict by:
-     *   - returning a {@code newCodec}, replaces the {@code oldCodec}
-     *   - returning a {@code oldCodec}, effectively ignores the {@code newCodec}
-     *   - returning a {@code null}, removes the {@code Codec} altogether
-     *
-     * By default, the implementation throws an exception that the codec with such name
-     * already exists.
-     *
-     * @param name codec name
-     * @param oldCodec existing codec
-     * @param newCodec new codec
-     * @return codec to use, or "null" for removal
-     */
-    default @Nullable Codec onConflict(String name, Codec oldCodec, Codec newCodec) {
-        throw new OpenSearchException("The codec with name " + name + " is already registered.");
-    }
 }